+++ source/renderer/app/components/wallet/widgets/VerticalSeparator.tsx
@@ -2,6 +2,6 @@ import React from 'react';
 
 import styles from './Divider.scss';
 
-export function Divider() {
+export function VerticalSeparator() {
   return <span className={styles.component} />;
 }