+++ source/renderer/app/stores/CurrencyStore.ts
@@ -7,6 +7,7 @@ import {
   getCurrencyFromCode,
 } from '../config/currencyConfig';
 import type { Currency, LocalizedCurrency } from '../types/currencyTypes';
+import { EventCategories } from '../analytics';
 
 export default class CurrencyStore extends Store {
   @observable
@@ -127,10 +128,23 @@ export default class CurrencyStore extends Store {
       this.getCurrencyRate();
       await this.api.localStorage.setCurrencySelected(selected.code);
     }
+
+    this.analytics.sendEvent(
+      EventCategories.SETTINGS,
+      'Changed currency',
+      code
+    );
   };
   @action
   _toggleCurrencyIsActive = () => {
     this.isActive = !this.isActive;
     this.api.localStorage.setCurrencyIsActive(this.isActive);
+
+    this.analytics.sendEvent(
+      EventCategories.SETTINGS,
+      `Turned ${
+        this.isActive ? 'on' : 'off'
+      } displaying ada balances in other currency`
+    );
   };
 }