+++ source/renderer/app/stores/SidebarStore.spec.ts
@@ -4,6 +4,7 @@ import type { ActionsMap } from '../actions/index';
 import { WalletSortBy, WalletSortOrder } from '../types/sidebarTypes';
 import type { SidebarWalletType } from '../types/sidebarTypes';
 import SidebarStore from './SidebarStore';
+import { noopAnalyticsTracker } from '../analytics';
 
 describe('Sidebar Store', () => {
   const api: Api = {
@@ -22,7 +23,7 @@ describe('Sidebar Store', () => {
       isLegacy?: boolean;
     }>;
   }) {
-    const sidebarStore = new SidebarStore(api, actions);
+    const sidebarStore = new SidebarStore(api, actions, noopAnalyticsTracker);
     sidebarStore.stores = {
       wallets: {
         all: wallets,